web-animations/animation-timing/animation-timing-direction.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationTiming direction attribute test</title>
<meta name="assert" content="The playback direction of the animation node as specified by one of the PlaybackDirection enumeration values">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationtiming-direction">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function createAnimationsToTest(test, timing) {
    var animation1 = new Animation(createDiv(test), KEYFRAMES, timing);
    var animation2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
    var animation3 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
    var animationGroup = new AnimationGroup([animation2], timing);
    var animationSequence = new AnimationSequence([animation3], timing);
    return [animation1, animation2, animation3];
}

async_test(function(t) {
    var timing = {
        duration: ANIMATION_DURATION
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(node.computedTiming.direction, 'normal',
            'Wrong AnimationNode.computedTiming.iterationStart value');

        player.ready.then(t.step_func(function() {
            player.currentTime = 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction');
            player.currentTime = ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');
            player.currentTime = ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');
            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that default value of AnimationTiming.direction is \'normal\' ' +
    'and animation is performed in forward direction');

async_test(function(t) {
    var timing = {
        direction: 'normal',
        duration: ANIMATION_DURATION,
        iterations: 3
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(node.computedTiming.direction, 'normal',
            'Wrong AnimationNode.computedTiming.iterationStart value');

        player.ready.then(t.step_func(function() {
            player.currentTime = 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction');
            player.currentTime = ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');
            player.currentTime = ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');

            player.currentTime = ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'even iterations');
            player.currentTime = ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction for even iterations');
            player.currentTime = ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'even iterations');
            player.currentTime =  2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'even iterations');

            player.currentTime = 2 * ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction for odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');

            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if AnimationTiming.direction value is \'normal\' then ' +
    'animation is performed in forward direction for all iterations');

async_test(function(t) {
    var timing = {
        direction: 'reverse',
        duration: ANIMATION_DURATION,
        iterations: 3
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(node.computedTiming.direction, 'reverse',
            'Wrong AnimationNode.computedTiming.iterationStart value');

        player.ready.then(t.step_func(function() {
            player.currentTime = 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction');
            player.currentTime = ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction');
            player.currentTime = ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction');

            player.currentTime = ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'even iterations');
            player.currentTime = ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction for even iterations');
            player.currentTime = ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'even iterations');
            player.currentTime =  2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'even iterations');

            player.currentTime = 2 * ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction for odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');

            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if AnimationTiming.direction value is \'reverse\' then ' +
    'animation is performed in reversed direction for all iterations');

async_test(function(t) {
    var timing = {
        direction: 'alternate',
        duration: ANIMATION_DURATION,
        iterations: 3
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(node.computedTiming.direction, 'alternate',
            'Wrong AnimationNode.computedTiming.iterationStart value');

        player.ready.then(t.step_func(function() {
            player.currentTime = 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction');
            player.currentTime = ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction');
            player.currentTime = ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction');

            player.currentTime = ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
                'Animation ' + type(node) + ' should be performed in normal direction for ' +
                'even iterations');
            player.currentTime = ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in ' +
                'normal direction for even iterations');
            player.currentTime = ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in normal direction for ' +
                'even iterations');
            player.currentTime =  2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in normal direction for ' +
                'even iterations');

            player.currentTime = 2 * ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction for odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'odd iterations');

            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if AnimationTiming.direction value is \'alternate\' then ' +
    'animation is performed in reversed direction for odd iterations and ' +
    'in normal directions for even iterations');

async_test(function(t) {
    var timing = {
        direction: 'alternate-reverse',
        duration: ANIMATION_DURATION,
        iterations: 3
    };
    var animations = createAnimationsToTest(t, timing);
    var doneCounter = 0;
    animations.forEach(function(animation) {
        var node = animation;
        if (animation.parent !== null) {
            node = node.parent;
        }
        var player = document.timeline.play(node);

        assert_equals(node.computedTiming.direction, 'alternate-reverse',
            'Wrong AnimationNode.computedTiming.iterationStart value');

        player.ready.then(t.step_func(function() {
            player.currentTime = 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction');
            player.currentTime = ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');
            player.currentTime = ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction');

            player.currentTime = ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_1,
                'Animation ' + type(node) + ' should be performed in forward reversed for ' +
                'even iterations');
            player.currentTime = ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in ' +
                'reversed direction for even iterations');
            player.currentTime = ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'even iterations');
            player.currentTime =  2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in reversed direction for ' +
                'even iterations');

            player.currentTime = 2 * ANIMATION_DURATION;
            assert_equals(window.getComputedStyle(animation.target).top, ANIMATION_TOP_0,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top, getExpectedTop(1),
                COMPUTED_STYLE_EPSILON, 'Animation ' + type(node) + ' should be performed in ' +
                'forward direction for odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION + ANIMATION_DURATION / 2;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION / 2), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');
            player.currentTime = 2 * ANIMATION_DURATION - 1;
            assert_approx_equals(window.getComputedStyle(animation.target).top,
                getExpectedTop(ANIMATION_DURATION - 1), COMPUTED_STYLE_EPSILON,
                'Animation ' + type(node) + ' should be performed in forward direction for ' +
                'odd iterations');

            doneCounter++;
            if (doneCounter === animations.length) {
                t.done();
            }
        }), t.unreached_func('Current ready promise should not be rejected'));
    });
}, 'Test that if AnimationTiming.direction value is \'alternate-reverse\' then ' +
    'animation is performed in normal direction for odd iterations and ' +
    'in reverse direction for even iterations');
</script>
</body>
